ARM: odroid-xu4: hack avoiding the invalid temperature by TMU broken
The temperatue value is weired when the TMU is broken from electronic
damage. In that case, XU4 is shutdown due to reaching the critical
temperature even if the real temperature is not high.

This chnage is to avoid invalid temperature using checking whether
the temperature value is valid or not. If the temperature value is
weired, then the temperature value is set to fake value.

Signed-off-by: Brian Kim <brian.kim@hardkernel.com>